+++ script/test/package.sh
@@ -4,21 +4,23 @@
 
 function is_server_up () {
     # check if server started
-    grep "serving" eva.txt
+    netstat -an | grep 0.0.0.0:5432
     return $?
 }
 
-eva_server &>> eva.txt &
+eva_server &> eva.txt &
+SERVER_PID=$!
 i=0
-while [ $i -lt 3 ];
+while [ $i -lt 5 ];
 do
+    echo "Waiting for server to launch, try $i"
     sleep 20
     is_server_up
     test_code=$?
     if [ $test_code == 0 ]; then
         break
     fi
-    i+=1
+    i=$((i+1))
 done
 
 echo "Contents of server log"
@@ -31,13 +33,24 @@ then
     exit "$test_code"
 fi
 
-eva_client &> client.txt &
-if [ "$test_code" -ne 0 ];
+cmd="exit"
+echo "$cmd"  | eva_client &> client.txt &
+
+# wait for client to launch
+sleep 5
+
+# shutdown server
+kill $SERVER_PID
+
+echo "Contents of client log"
+cat client.txt
+
+grep "failed" client.txt
+if [ "$?" -ne 1 ];
 then
     echo "Client did not start"
     echo "$test_code"
     exit "$test_code"
 fi
 
-head -n20 client.txt
 exit 0