+++ test/integration_tests/test_open.py
@@ -30,7 +30,6 @@ class OpenTests(unittest.TestCase):
     def setUp(self):
         CatalogManager().reset()
         ConfigurationManager()
-
         # Load built-in UDFs.
         load_inbuilt_udfs()